remove all the old txsponsor stuff
