+++ wrangler/contracts/directory.md
@@ -53,9 +53,3 @@ SDAODX4C4XK3P5UMRPYHVIF57E5PAJOLSP73DIAXJ4WDJ4XGQ7XL3BQO
 GAV277VQFXN4AWABKHNMERLQ3ZC4W3COSGO4FYL23PTYFUZT27ABMQSF
 SAAOIFVIA3SUG56LBVN5Y4DGRFRLQO7M3OUCQXFHY7E7Q5XCLWKCQVC3
 ```
-
-## Sponsor
-```
-GDTSBG5TME7QYC2CVGR5LIWKA4LUJAWWBDFRRXIPAWJQQNUBUETTQKNQ
-SAX3WMY4GEFSMSCVUBGPX6NNP3YY6HPULASNIAXTGPCLX6LA4NF2JZSU
-```
