+++ wrangler/wrangler.toml
@@ -7,6 +7,5 @@ kv_namespaces = [
    { binding = "META", id = "239e6a0c4c8d47569eb202278fd99e58" },
    { binding = "TX_FUNCTIONS", id = "bb7ba59b38e040bbab7d71dfdb826911" },
    { binding = "TX_FEES", id = "bc666162fd1c44f0b22c605e14c07597" },
-   { binding = "TX_SPONSORS", id = "01269d925c44431aaab92c31307e8599" }
 ]
-vars = { STELLAR_NETWORK = "TESTNET", HORIZON_URL = "https://horizon-testnet.stellar.org", TURRET_ADDRESS = "GB4OYM7TQTJSROWXHOJLKAX2IJ2QN4I6S6GCJH4MGWVTAO5Q5DPNADXX", TURRET_RUN_URL = "https://ylorcf18ii.execute-api.us-east-1.amazonaws.com", XLM_FEE_MIN = 1, XLM_FEE_MAX = 10, UPLOAD_DIVISOR = 1000, RUN_DIVISOR = 100000 }
+vars = { STELLAR_NETWORK = "TESTNET", HORIZON_URL = "https://horizon-testnet.stellar.org", TURRET_ADDRESS = "GB4OYM7TQTJSROWXHOJLKAX2IJ2QN4I6S6GCJH4MGWVTAO5Q5DPNADXX", TURRET_RUN_URL = "https://ylorcf18ii.execute-api.us-east-1.amazonaws.com", TX_FUNCTION_FEE_DAYS_TTL = 180, XLM_FEE_MIN = 1, XLM_FEE_MAX = 10, UPLOAD_DIVISOR = 1000, RUN_DIVISOR = 100000 }
